+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Break after reset under certain circumstances.
+
 ## [0.1.2] - 2018-05-30
 
 ### Fixed